fix(activerecord/encryption): move outer Cipher class to encryption/cipher.ts, fix name collision
